Adding message queue

+++ src/com/PendencySystem/service/impl/PendencyManagementServiceImpl.java
@@ -1,7 +1,5 @@
 package com.PendencySystem.service.impl;
 
-import com.PendencySystem.exceptions.NotFoundException;
-import com.PendencySystem.model.Tag;
 import com.PendencySystem.model.Transaction;
 import com.PendencySystem.repository.interfaces.TagRepository;
 import com.PendencySystem.repository.interfaces.TransactionRepository;
@@ -14,33 +12,36 @@ public class PendencyManagementServiceImpl implements PendencyManagementService 
     private final TransactionRepository transactionRepository;
     private final TagRepository tagRepository;
 
-    public PendencyManagementServiceImpl(TransactionRepository transactionRepository, TagRepository tagRepository){
-        this.tagRepository = tagRepository;
+    public PendencyManagementServiceImpl(TransactionRepository transactionRepository,
+                                         TagRepository tagRepository) {
         this.transactionRepository = transactionRepository;
+        this.tagRepository = tagRepository;
     }
 
     @Override
     public void startTracking(Integer id, List<String> hierarchicalTags) {
+        if (id == null || hierarchicalTags == null || hierarchicalTags.isEmpty()) {
+            throw new IllegalArgumentException("ID and tags are required");
+        }
         transactionRepository.create(id, hierarchicalTags);
-        tagRepository.createTags(hierarchicalTags);
+        tagRepository.incrementTags(hierarchicalTags);
     }
 
     @Override
     public void stopTracking(Integer id) {
+        if (id == null) {
+            throw new IllegalArgumentException("ID is required");
+        }
         Transaction transaction = transactionRepository.get(id);
-        List<String> hierarchicalTags = transaction.getTags();
-        tagRepository.deleteTransactionFromTags(hierarchicalTags);
+        tagRepository.decrementTags(transaction.getTags());
         transactionRepository.remove(id);
-
     }
 
     @Override
-    public Integer getCounts(List<String> tags) {
-        try{
-            Tag leafTag = tagRepository.getLeafTag(tags);
-            return leafTag.getCount();
-        }catch (NotFoundException exception){
-            return 0;
+    public int getCounts(List<String> tags) {
+        if (tags == null || tags.isEmpty()) {
+            throw new IllegalArgumentException("Tags are required");
         }
+        return tagRepository.getCounts(tags);
     }
 }
